Run `live` on the MainActor, allow multiple `delegate` streams
When testing the current (Swift Concurrency) approach, the delegate
method calls were not being invoked, and after some investigation it
seems the problem was related with the `RunLoop` in which the
`CLLocationManager` was created. With the latest `Dependencies`/`TCA`
implementations, dependencies' access isn't bound to the `@MainActor`,
so code is run from arbitrary queues/threads spanning multiple RunLoops.

According to `CLLocationManager`'s
[documentation](https://developer.apple.com/documentation/corelocation/cllocationmanager):

> "Core Location calls the methods of your delegate object using the
> `RunLoop` of the thread on which you initialized the
> `CLLocationManager` object.

By forcing the `LocationManager.live`'s `CLLocationManager` instance to
be instantiated on the main `RunLoop` via `@MainActor`, the delegate
methods are correctly invoked. To minimize problems, all `.live`
instance APIs' closures were annotated with `@MainActor`.

Furthermore, as the `Dependencies` instances are shared and can be
called from multiple places concurrently, the `LocationManagerDelegate`
was updated to allow forwarding delegate calls to multiple streams
provided via `delegate()`.

Ran `swift-format` and performed some cleanups.